Optimizations, migration away from DotSDL to plain SDL3 (#70)
* Migrate Exits to use refs

* Move more to plain ol' SDL3

* Fresh new AudioPresenter implementation

* Remove DotSDL dependency

* Remove the last remaining heap allocation in Engine

* Momma, that's a fast glyph blitter

* Set up vsync on video render, add comments, mandate 2x audio buffers

* Use a lock on the audio buffer

* DotSDL is removed

* Unlink DotSDL from submodules

+++ Source/Lyon/Presenters/IAudioPresenter.cs
@@ -4,7 +4,7 @@ namespace Lyon.Presenters;
 
 public interface IAudioPresenter
 {
-    double Volume { get; set; }
+    float Volume { get; set; }
     void Start();
     void Stop();
     void Update(AudioComposerDataEventArgs buffer);
